Weight changed from double to int to ease the comparisons

+++ graph.h
@@ -25,25 +25,25 @@ class Graph
 		Graph();
 		~Graph();
 
-		typedef map<string, map<string, double>> AdjMatrix;
-		typedef AdjMatrix::iterator r_iter; // iterator for adjacency matrix rows
-		typedef map<string, double>::iterator c_iter; // iterator for adjacency matrix columns
+	typedef map<string, map<string, int>> AdjMatrix;
+	typedef AdjMatrix::iterator r_iter; // iterator for adjacency matrix rows
+	typedef map<string, int>::iterator c_iter; // iterator for adjacency matrix columns
 
 		void addVertex(const string name);
 		void removeVertex(const string name);
 
-		void addEdge(const string v1, const string v2, const double weight);
-		void removeEdge(const string v1, const string v2);
-		bool edgeExist(const string v1, const string v2);
-		double getWeight(const string v1, const string v2);
-		void setWeight(const string v1, const string v2, const double weight);
+	void addEdge(const string v1, const string v2, const int weight);
+	void removeEdge(const string v1, const string v2);
+	bool edgeExist(const string v1, const string v2);
+	int getWeight(const string v1, const string v2);
+	void setWeight(const string v1, const string v2, const int weight);
 
 		int countVertices() const;
 		int countEdges() const;
 
 		void print() const;
 	private:
-		constexpr static const double INF = numeric_limits<double>::max();
+		constexpr static const int INF = numeric_limits<int>::max();
 		int edges;
 		AdjMatrix adjMatrix;
 };
@@ -109,7 +109,7 @@ int Graph::countEdges() const
 	return edges;
 }
 
-void Graph::addEdge(const string v1, const string v2, const double weight)
+void Graph::addEdge(const string v1, const string v2, const int weight)
 {
 	if(v1 == v2)
         {
@@ -148,12 +148,12 @@ bool Graph::edgeExist(const string v1, const string v2)
 	return adjMatrix[v1][v2] != INF;
 }
 
-double Graph::getWeight(const string v1, const string v2)
+int Graph::getWeight(const string v1, const string v2)
 {
 	return adjMatrix[v1][v2];
 }
 
-void Graph::setWeight(const string v1, const string v2, const double weight)
+void Graph::setWeight(const string v1, const string v2, const int weight)
 {
 	if (v1 == v2) throw GraphExcept("The vertex can't have an edge with itself");
 	adjMatrix[v1][v2] = weight;
